Finished repacking the Rserve server
The server now maintains a single link to Rserve.  This was done in
order to avoid potential hang time associated with a request arriving
for execution on one R process while the server is busy processing a
response from another R process.

Hopefully this will simplify things by allowing the logic associated
with multiple links to different R sessions to be moved into the
Supervisor.